ex: 4.15-4.23*

+++ part4/controllers/blogs.js
@@ -1,11 +1,14 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const middleware = require('../utils/middleware')
 
 
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-  return response.json(blogs)
+  const blogs = await Blog.find({}).populate('user',{username: 1, name: 1})
+  response.json(blogs)
   // Blog
   //   .find({})
   //   .then(blogs => {
@@ -13,13 +16,40 @@ blogsRouter.get('/', async (request, response) => {
   //   })
 })
 
-blogsRouter.post('/', async (request, response,next) => {
-  const blog = new Blog(request.body)
+// const getTokenFrom = request => {
+//   const authorization = request.get('authorization')
+//   if(authorization && authorization.startsWith(`Bearer `)){
+//     return authorization.replace(`Bearer `,'')
+//   }
+//   return null
+// }
+
+blogsRouter.post('/',middleware.userExtractor, async (request, response,next) => {
+  try{
+    const body = request.body
+  
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  // if (!decodedToken.id){
+  //   return response.status(401).json({error: 'token invalid'})
+  // }
+
+  const user = request.user
+  
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.author,
+    likes: body.likes,
+    user: user._id
+  })
   if (!blog.title || ! blog.url){
     response.status(400).end()
   }
-  try{
+
   const result = await blog.save()
+  user.blogs = user.blogs.concat(result._id)
+  await user.save()
+
   response.status(201).json(result)
 } catch (ex){
   next(ex)
@@ -32,10 +62,22 @@ blogsRouter.post('/', async (request, response,next) => {
   //   .catch(error => next(error))
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id',middleware.userExtractor, async (request, response, next) => {
   try{
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    // const decodedToken = jwt.verify(request.token,process.env.SECRET)
+    // if (!decodedToken.id){
+    //   return response.status(401).json({error: 'token invalid'})
+    // }
+    // const user = await User.findById(decodedToken.id)
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+    if(blog.user.toString() === user._id.toString()){
+      await Blog.findByIdAndDelete(request.params.id)
+      response.status(204).end()
+    } else{
+      return response.status(403).json({error: 'only the creater and delete this blog'})
+    }
+    
   } catch(ex){
     next(ex)
   }
